Refactor: service 로직 분리
